requirements.txt
nfcpy
requests
RPi.GPIO
